Pull TickingCassetteListener into its own file
